% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwByPointsFast.R
\name{gwByPoints}
\alias{gwByPoints}
\title{Geographical Weights PCA by points}
\usage{
gwByPoints(coorXY, pointXY, env, taus)
}
\arguments{
\item{coorXY}{Matrix longtiude/latitude.}

\item{pointXY}{matrix by points longitude/latitude}

\item{env}{Environemntal layers}

\item{taus}{Factor Tau for distance matrix}
}
\value{
Returns the geographical weights distance among environmental layers by points
}
\description{
This function perform by points
}
