% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwCovariance.R
\name{gwCovariance}
\alias{gwCovariance}
\title{Geographical weights Covariance - Matrices}
\usage{
gwCovariance(y, ww)
}
\arguments{
\item{y}{matrix. m by n matrix.}

\item{ww}{Geographical Weights Distances}
}
\value{
`gwCovariance` Returns the weighted covariance matrix of y (by columns)
}
\description{
gwCovariance compute the variance of x and the covariance or correlation of x and y if these are vectors.
}
