% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_distance.R
\name{gwDistance}
\alias{gwDistance}
\title{Trapezoidal integration for Partial ROC curve}
\usage{
gwDistance(coodXY, pointXY, tau)
}
\arguments{
\item{coodXY}{fractional area predicted presence}

\item{pointXY}{sensibility at threshold (fract_area value).}

\item{tau}{sadasda}
}
\value{
Returns the area under the Partial ROC curve.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
