% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwEigen.R
\name{gwEigen}
\alias{gwEigen}
\title{EigenValues}
\usage{
gwEigen(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
`Eigen` rounds the eigen values
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
