% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwHeterogen.R
\name{gwHeterogen}
\alias{gwHeterogen}
\title{Geographical Weights Principal Component Analysis}
\usage{
gwHeterogen(
  layers,
  baseRaster,
  comp_each = 1000,
  n_cores = 15,
  taus = 0.01,
  parallel = TRUE,
  save = TRUE,
  outdir = "Output_hetero",
  outname = "heterog",
  rmpi = FALSE
)
}
\arguments{
\item{layers}{Rast terra-object,}

\item{baseRaster}{Raster-object. Base Raster object to write output}

\item{comp_each}{Comp_by_range.}

\item{n_cores}{Number of cores}

\item{taus}{Factor Tau for distance matrix}

\item{parallel}{boolean. `TRUE` run with parallel. Run only parallel. FALSE, doesn't works}

\item{save}{boolean. `TRUE` save raster file in local computer}

\item{outdir}{character. Folder name}

\item{outname}{Full name of rasterfile}

\item{rmpi}{default FALSE. Use RMPI for cluster computer (High Performance Computer)}
}
\value{
Returns the geographical weights distance among environmental layers or variables.
}
\description{
This function perform a environmental heterogeneity on spatial space.
}
