% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwHeterogeneity.R
\name{gwHeterogeneity}
\alias{gwHeterogeneity}
\title{Geographical Weights Principal Component Analysis}
\usage{
gwHeterogeneity(
  layers,
  factor,
  parallel = TRUE,
  cores = 4,
  save = FALSE,
  outdir = "."
)
}
\arguments{
\item{layers}{RasterStack object or Rast object.}

\item{factor}{Numeric. Factor by moving  windows}

\item{parallel}{boolean. `TRUE` run with parallel.}

\item{cores}{Numeric. cores or threats}

\item{save}{boolean. `TRUE` save raster file in local computer}

\item{outdir}{character. Folder name}
}
\value{
Returns the geographical weights distance among environmental layers or variables.
}
\description{
This function computes and returns the distance matrix computed by using the specified distance measure to compute the distances between the rows of a data matrix.
}
