% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwHeter_rmpi.R
\name{gwheterroRmpi}
\alias{gwheterroRmpi}
\title{Heterovalues mrpi}
\usage{
gwheterroRmpi(coodXY, pointXY, Y, tau)
}
\arguments{
\item{coodXY}{a numeric vector.}

\item{pointXY}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed (see ‘Details’)}

\item{Y}{sadasd as}

\item{tau}{adadsa}
}
\value{
`round` rounds the values in its first argument to the specified number of decimal places (default 0). See ‘Details’ about “round to even” when rounding off a 10.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
