% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwMutiMatrix.R
\name{gwMultiMatrix}
\alias{gwMultiMatrix}
\title{Trapezoidal integration for Partial ROC curve}
\usage{
gwMultiMatrix(A, B)
}
\arguments{
\item{A}{fractional area predicted presence}

\item{B}{sensibility at threshold (fract_area value).}
}
\value{
Returns the area under the Partial ROC curve.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
