% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwRound.R
\name{gwRound}
\alias{gwRound}
\title{Rounding of Numbers}
\usage{
gwRound(A, digits = 10L)
}
\arguments{
\item{A}{a numeric vector.}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed (see ‘Details’)}
}
\value{
`round` rounds the values in its first argument to the specified number of decimal places (default 0). See ‘Details’ about “round to even” when rounding off a 10.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
