% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwTraspose.R
\name{gwTraspose}
\alias{gwTraspose}
\title{Trapezoidal integration for Partial ROC curve}
\usage{
gwTraspose(m, byrow = FALSE)
}
\arguments{
\item{m}{fractional area predicted presence}

\item{byrow}{boolean asdad}
}
\value{
Returns the area under the Partial ROC curve.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
