% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallc.R
\name{MannKendallc}
\alias{MannKendallc}
\title{Mann-Kendall Test Corrected}
\usage{
MannKendallc(x)
}
\arguments{
\item{vc}{Time series data vector.}
}
\value{
Returns the geographical weights distance among environmental layers or variables.
}
\description{
Time series data is often influenced by serial-correlation. 
When data is not random and influenced by auto-correlation, 
Modified Mann-Kendall tests are to be used in trend detection. 
Yue and Wang (2004) have proposed variance correction approach to address the issue of serial correlation in Trend analysis. 
Trend is removed from the series and effective sample size is calulated using significant serial correlation coefficients.
}
\references{
Mann, H. B. (1945). Nonparametric Tests Against Trend. Econometrica, 13(3), 245–259. http://doi.org/10.1017/CBO9781107415324.00

Kendall, M. (1975). Multivariate analysis. Charles Griffin. Londres. 0-85264-234-2.

Sen, P. K. (1968). Estimates of the Regression Coefficient Based on Kendall’s Tau. Journal of the American Statistical Association, 63(324), 1379. http://doi.org/10.2307/2285891

Yue, S., & Wang, C. Y. (2004). The Mann-Kendall test modified by effective sample size to detect trend in serially correlated hydrological series. Water Resources Management, 18(3), 201–218. http://doi.org/10.1023/B:WARM.0000043140.61082.60
}
