% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwDistance.R
\name{gwDistance}
\alias{gwDistance}
\title{Geographical Weights Distance Matrix Computation}
\usage{
gwDistance(coodXY, pointXY, tau)
}
\arguments{
\item{coodXY}{Matrix. Matrix of coordinates e.g. longitude/latitude, unit=long/lat}

\item{pointXY}{Matrix. Point pairs e.g. longitude/latitude}

\item{tau}{sadasda}
}
\value{
Returns the geographical weights distance among environmental layers or variables.
}
\description{
This function computes and returns the distance matrix computed by using the specified distance measure to compute the distances between the rows of a data matrix.
}
