% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwMutiMatrix.R
\name{gwMultiMatrix}
\alias{gwMultiMatrix}
\title{Matrix Multiplication}
\usage{
gwMultiMatrix(A, B)
}
\arguments{
\item{A}{numeric or complex matrices or vectors.}

\item{B}{numeric or complex matrices or vectors.}
}
\value{
Multiplies two matrices
}
\description{
Multiplies two matrices, if they are conformable. If one argument is a vector, it will be promoted to either a row or column matrix to make the two arguments conformable. If both are vectors of the same length, it will return the inner product (as a matrix).
}
