% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwRound.R
\name{gwRound}
\alias{gwRound}
\title{Trapezoidal integration for Partial ROC curve}
\usage{
gwRound(A, digits = 10L)
}
\arguments{
\item{A}{fractional area predicted presence}

\item{digits}{diitsos}
}
\value{
Returns the area under the Partial ROC curve.
}
\description{
trapozoid_roc applies the trapezoidal integration for Partial ROC curve.
}
