% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwTraspose.R
\name{gwTraspose}
\alias{gwTraspose}
\title{Matrix Transpose}
\usage{
gwTraspose(m, byrow = FALSE)
}
\arguments{
\item{m}{a matrix or data.frame, typically.}

\item{byrow}{a boolean. Ordering by colums `default=FALSE`. `TRUE` order by rows}
}
\value{
A matrix, with dim and dimnames constructed appropriately from those of `m`, 
and other attributes except names copied across.
}
\description{
Given a matrix or data.frame `m`, `gwTraspose` returns the transpose of `m`.
}
\examples{
# matrix
a <- matrix(1:28, 4, 7)
gwTraspose(a, byrow=TRUE)

}
